+++ src/ansys/mapdl/core/mapdl_core.py
@@ -61,10 +61,10 @@ from ansys.mapdl.core.errors import (
     MapdlInvalidRoutineError,
     MapdlRuntimeError,
 )
+from ansys.mapdl.core.information import Information
 from ansys.mapdl.core.inline_functions import Query
 from ansys.mapdl.core.mapdl_types import MapdlFloat
 from ansys.mapdl.core.misc import (
-    Information,
     check_valid_routine,
     last_created,
     random_string,
