+++ tests/test_misc.py
@@ -21,7 +21,6 @@
 # SOFTWARE.
 
 """Small or misc tests that don't fit in other test modules"""
-import inspect
 import os
 import pathlib
 
@@ -145,48 +144,6 @@ def test_no_return(mapdl, cleared):
     assert fun(mapdl) is None
     last_keypoint = np.array(mapdl.klist().splitlines()[-1].split(), dtype=float)[0:4]
     assert np.allclose(last_keypoint, np.array([1, 1, 1, 1]))
-
-
-def test_mapdl_info(mapdl, capfd):
-    info = mapdl.info
-    for attr, value in inspect.getmembers(info):
-        if not attr.startswith("_") and attr not in ["title", "stitles"]:
-            assert isinstance(value, str)
-
-            with pytest.raises(AttributeError):
-                setattr(info, attr, "any_value")
-
-    assert "PyMAPDL" in mapdl.info.__repr__()
-    out = info.__str__()
-
-    assert "ansys" in out.lower()
-    assert "Product" in out
-    assert "MAPDL Version" in out
-    assert "UPDATE" in out
-
-
-def test_info_title(mapdl):
-    title = "this is my title"
-    mapdl.info.title = title
-    assert title == mapdl.info.title
-
-
-def test_info_stitle(mapdl):
-    info = mapdl.info
-
-    assert all([not each for each in info.stitles])
-    stitles = ["asfd", "qwer", "zxcv", "jkl"]
-    info.stitles = "\n".join(stitles)
-
-    assert stitles == info.stitles
-
-    stitles = stitles[::-1]
-
-    info.stitles = stitles
-    assert stitles == info.stitles
-
-    info.stitles = None
-    assert all([not each for each in info.stitles])
 
 
 @pytest.mark.parametrize("file_", ["dummy.dumdum", "dumdum.dummy"])
